+++ packages/terra-avatar/src/terra-dev-site/test/avatar/avatar/UpdateAvatar.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Avatar from '../../../../index';
-import exampleAvatarImage from '../../../assets/150x150.jpg';
+import exampleAvatarImage from '../../../assets/163x40.png';
 import exampleAvatarImage2 from '../../../assets/200x133.jpg';
 import exampleAvatarImage3 from '../../../assets/170x251.jpg';