+++ packages/terra-avatar/src/variants/Avatar.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
-import objectFitImages from 'object-fit-images'; // Adding polyfill for IE.
 import styles from '../common/Avatar.module.scss';
 import {
   AVATAR_VARIANTS, generateImagePlaceholder, generateImage, setColor,
@@ -65,10 +64,6 @@ class Avatar extends React.Component {
     };
 
     this.handleFallback = this.handleFallback.bind(this);
-  }
-
-  componentDidMount() {
-    objectFitImages();
   }
 
   handleFallback() {